Split the tui part into its own module.
